prv/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "airflow-provider-remsvc"
version = "1.0.0"
description = "Apache Airflow provider for RemSvc remote execution via gRPC"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
dependencies = [
    # NOTE: apache-airflow is intentionally NOT listed here as a hard
    # dependency — it is provided by the target Airflow environment.
    # Add it only in the [dev] extras below.
    "grpcio>=1.51.0",
    "grpcio-status>=1.51.0",
    "protobuf>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "apache-airflow>=2.2.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

# Airflow discovers the provider via this entry point
[project.entry-points."apache_airflow_provider"]
remsvc = "remsvc_provider"

[tool.hatch.build.targets.wheel]
# Only package remsvc_provider/ — remsvc_proto/ is generated separately
packages = ["remsvc_provider"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
